feat: redirect data.forgejo.org to code.forgejo.org
code.forgejo.org currently host a number of assets that need to be highly
available:

- most actions (e.g. https://code.forgejo.org/actions)
- OCI mirror (e.g. https://code.forgejo.org/oci/)

a growing number of Forgejo CI workflows and third party repository
depends on them.

Their availability could be improved if they were moved to Forgejo
instances that are read-only. That requires a change of the URL used to
access them from code.forgejo.org (read-write) to
data.forgejo.org (read-only).

This pull request is a first step in this direction: the redirect allows
to use data.forgejo.org instead of code.forgejo.org in workflows or
other places where actions and container images are needed.

It will make it possible to use data.forgejo.org as a default for actions
instead of code.forgejo.org as soon as v10 is published.